+++ src/main.tsx
@@ -10,46 +10,13 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 import { ThemeProvider } from "styled-components";
 import { theme as antdTheme, ConfigProvider } from "antd";
+import { theme } from "./theme.tsx";
 
 const { useToken } = antdTheme;
 
 const queryClient = new QueryClient();
 
-export const theme = {
-  token: {
-    colorPrimary: "#05409A",
-    fontFamily: "Roboto, sans-serif",
-    borderRadius: 2,
-    colorBgContainer: "#FFFFFF",
-    colorBgBase: "#FFFFFF",
-    colorTextLightSolid: "#ffffff",
-    colorError: "#DB001B",
-    colorBgLayout: "#FFFFFF",
-    colorTextDescription: "#161718",
-    fontSize: 16,
-    fontWeightStrong: 700,
-    fontWeight: 700,
-  },
-  components: {
-    Layout: {
-      headerBg: "#FFFFFF",
-    },
-    Table: {
-      headerBg: "#EBEBED",
-      headerSplitColor: "transparent",
-      rowBg: "#FFFFFF", 
-      rowSelectedBg: "#F6F6F6", 
-    },
-    Pagination: {
-      colorPrimary: "#ffffff",
-      itemActiveBg: "#05409A",      
-    },
-    Button: {
-      fontWeight: 700,
-      
-    },
-  },
-};
+ 
 
 function ThemedApp() {
   const { token } = useToken();
