+++ src/theme.tsx
@@ -1,0 +1,37 @@
+ 
+ 
+
+export const theme = {
+  token: {
+    colorPrimary: "#05409A",
+    fontFamily: "Roboto, sans-serif",
+    borderRadius: 2,
+    colorBgContainer: "#FFFFFF",
+    colorBgBase: "#FFFFFF",
+    colorTextLightSolid: "#ffffff",
+    colorError: "#DB001B",
+    colorBgLayout: "#FFFFFF",
+    colorTextDescription: "#161718",
+    fontSize: 16,
+    fontWeightStrong: 700,
+    fontWeight: 700,
+  },
+  components: {
+    Layout: {
+      headerBg: "#FFFFFF",
+    },
+    Table: {
+      headerBg: "#EBEBED",
+      headerSplitColor: "transparent",
+      rowBg: "#FFFFFF", 
+      rowSelectedBg: "#F6F6F6", 
+    },
+    Pagination: {
+      colorPrimary: "#ffffff",
+      itemActiveBg: "#05409A",      
+    },
+    Button: {
+      fontWeight: 700,      
+    },
+  },
+};
